+++ build_runner/lib/src/generate/phase.dart
@@ -12,7 +12,7 @@ import 'input_matcher.dart';
 
 /// A "phase" in the build graph, which represents running a one or more
 /// builders on a set of sources.
-abstract class BuildAction {
+abstract class BuildPhase {
   /// Whether to run lazily when an output is read.
   ///
   /// An optional build action will only run if one of its outputs is read by
@@ -37,10 +37,11 @@ abstract class BuildAction {
   int get identity;
 }
 
-/// A [BuildAction] that uses a single [Builder] to generate files.
-class BuilderBuildAction extends BuildAction {
+/// A [BuildPhase] that uses a single [Builder] to generate files.
+class InBuildPhase extends BuildPhase {
   final Builder builder;
 
+  final String builderLabel;
   final BuilderOptions builderOptions;
   final InputMatcher generateFor;
   final String package;
@@ -51,15 +52,16 @@ class BuilderBuildAction extends BuildAction {
   @override
   final bool hideOutput;
 
-  BuilderBuildAction._(this.package, this.builder, this.builderOptions,
+  InBuildPhase._(this.package, this.builder, this.builderOptions,
       {@required this.targetSources,
       @required this.generateFor,
+      @required this.builderLabel,
       bool isOptional,
       bool hideOutput})
       : this.isOptional = isOptional ?? false,
         this.hideOutput = hideOutput ?? false;
 
-  /// Creates an [BuildAction] for a normal [Builder].
+  /// Creates an [BuildPhase] for a normal [Builder].
   ///
   /// The build target is defined by [package] as well as [targetSources]. By
   /// default all sources in the target are used as primary inputs to the
@@ -70,9 +72,10 @@ class BuilderBuildAction extends BuildAction {
   ///
   /// [hideOutput] specifies that the generated asses should be placed in the
   /// build cache rather than the source tree.
-  factory BuilderBuildAction(
+  factory InBuildPhase(
     Builder builder,
     String package, {
+    String builderKey,
     InputSet targetSources,
     InputSet generateFor,
     BuilderOptions builderOptions,
@@ -83,9 +86,12 @@ class BuilderBuildAction extends BuildAction {
         new InputMatcher(targetSources ?? const InputSet());
     var generateForMatcher = new InputMatcher(generateFor ?? const InputSet());
     builderOptions ??= const BuilderOptions(const {});
-    return new BuilderBuildAction._(package, builder, builderOptions,
+    return new InBuildPhase._(package, builder, builderOptions,
         targetSources: targetSourceMatcher,
         generateFor: generateForMatcher,
+        builderLabel: builderKey == null || builderKey.isEmpty
+            ? _builderLabel(builder)
+            : _simpleBuilderKey(builderKey),
         isOptional: isOptional,
         hideOutput: hideOutput);
   }
@@ -111,20 +117,20 @@ class BuilderBuildAction extends BuildAction {
       ]);
 }
 
-/// A [BuildAction] that can run multiple [PostProcessBuilderAction]s to
+/// A [BuildPhase] that can run multiple [PostBuildAction]s to
 /// generate files.
 ///
 /// There should only be one of these per build, and it should be the final
 /// phase.
-class PostProcessBuildAction extends BuildAction {
-  final List<PostProcessBuilderAction> builderActions;
+class PostBuildPhase extends BuildPhase {
+  final List<PostBuildAction> builderActions;
 
   @override
   bool get hideOutput => true;
   @override
   bool get isOptional => false;
 
-  PostProcessBuildAction(this.builderActions);
+  PostBuildPhase(this.builderActions);
 
   @override
   String toString() =>
@@ -139,20 +145,25 @@ class PostProcessBuildAction extends BuildAction {
         ]));
 }
 
-/// Part of a larger [PostProcessBuildAction], applies a single
+/// Part of a larger [PostBuildPhase], applies a single
 /// [PostProcessBuilder] to a single [package] with some additional options.
-class PostProcessBuilderAction {
+class PostBuildAction {
   final PostProcessBuilder builder;
+  final String builderLabel;
   final BuilderOptions builderOptions;
   final InputMatcher generateFor;
   final String package;
   final InputMatcher targetSources;
 
-  PostProcessBuilderAction(this.builder, this.package,
-      {@required BuilderOptions builderOptions,
+  PostBuildAction(this.builder, this.package,
+      {String builderKey,
+      @required BuilderOptions builderOptions,
       @required InputSet targetSources,
       @required InputSet generateFor})
-      : builderOptions = builderOptions ?? const BuilderOptions(const {}),
+      : builderLabel = builderKey == null || builderKey.isEmpty
+            ? _builderLabel(builder)
+            : _simpleBuilderKey(builderKey),
+        builderOptions = builderOptions ?? const BuilderOptions(const {}),
         targetSources = new InputMatcher(targetSources ?? const InputSet()),
         generateFor = new InputMatcher(generateFor ?? const InputSet());
 
@@ -162,6 +173,23 @@ class PostProcessBuilderAction {
         package,
         targetSources,
       ]);
+}
+
+/// If we have no key find a human friendly name for the Builder.
+String _builderLabel(Object builder) {
+  var label = '$builder';
+  if (label.startsWith('Instance of \'')) {
+    label = label.substring(13, label.length - 1);
+  }
+  return label;
+}
+
+/// Change "angular|angular" to "angular".
+String _simpleBuilderKey(String builderKey) {
+  if (!builderKey.contains('|')) return builderKey;
+  var parts = builderKey.split('|');
+  if (parts[0] == parts[1]) return parts[0];
+  return builderKey;
 }
 
 final _deepEquals = const DeepCollectionEquality();